backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pyotp==2.9.0
qrcode[pil]==7.4.2
pillow==10.3.0
twilio==9.0.4
apscheduler==3.10.4
playwright==1.44.0
beautifulsoup4==4.12.3
aiohttp==3.9.5
redis==5.0.4
celery==5.4.0
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
